Move AbstractWires
