+++ pennylane/wires.py
@@ -739,52 +739,6 @@ class Wires(Sequence):
         """Right-hand version of __xor__."""
         return Wires(set(_process(other)) ^ set(self.labels))
 
-    def __class_getitem__(cls, item) -> "AbstractWires":
-        if not isinstance(item, int) and item != ...:
-            raise TypeError(
-                f"AbstractWires can only be subscripted with integers and Ellipsis. Got {item}."
-            )
-        return AbstractWires(item)
-
-
-@dataclass(frozen=True)
-class AbstractWires:
-    """An abstract representation of a sequence of wires that contains the number
-    of wires, useful for resource calculations.
-
-    Args:
-        num_wires (int): The number of wires
-    """
-
-    num_wires: int | types.EllipsisType
-
-    def __eq__(self, other) -> bool:
-        if isinstance(other, AbstractWires):
-            return self.num_wires == other.num_wires
-
-        raise TypeError("Tried to check equality against an abstract wire register.")
-
-    @property
-    def shape(self) -> tuple[int]:
-        """The number of wires expressed as shape ``(num_wires, )``."""
-        return (self.num_wires,)
-
-    @property
-    def dtype(self):
-        """np.int64.  The dtype of wires when used with Catalyst."""
-        return np.int64
-
-    def __hash__(self):
-        return hash(("AbstractWires", self.num_wires))
-
-    def __len__(self) -> int:
-        return self.num_wires
-
-    def __instancecheck__(self, instance):
-        if not isinstance(instance, Wires):
-            return False
-        return len(instance) == self.num_wires
-
 
 WiresLike = Wires | Iterable[Hashable] | Hashable
