+++ tests/test_wires.py
@@ -22,7 +22,8 @@ import pytest
 
 import pennylane as qp
 from pennylane.exceptions import WireError
-from pennylane.wires import AbstractWires, Wires
+from pennylane.typing import AbstractWires
+from pennylane.wires import Wires
 
 if util.find_spec("jax") is not None:
     jax = import_module("jax")
